Fix random int registration

+++ SCP-575/SCP-575.cs
@@ -84,12 +84,11 @@ namespace SCP575
 			this.AddConfig(new ConfigSetting("575_keter_kill", false, SettingType.BOOL, true, "If SCP-575's keter event should kill players instead of damage them."));
 			this.AddConfig(new ConfigSetting("575_keter_kill_num", 1, SettingType.NUMERIC, true, "The number of players killed during timed Keter events.));"));
 			this.AddConfig(new ConfigSetting("575_random_events", false, SettingType.BOOL, true, "If 575 events should be randomized."));
-			this.AddConfig(new ConfigSetting("575_random_min", 60, SettingType.FLOAT, true, "The minimum amount of time between random events."));
-			this.AddConfig(new ConfigSetting("575_random_max", 300, SettingType.FLOAT, true, "The maximum wait time between random events."));
-			this.AddConfig(new ConfigSetting("575_random_dur_min", 11, SettingType.FLOAT, true, "The minimum amount of time a random event should last."));
-			this.AddConfig(new ConfigSetting("575_random_dur_max", 90, SettingType.FLOAT, true, "The maximum amount of time a random event should last."));
+			this.AddConfig(new ConfigSetting("575_random_min", 60, SettingType.NUMERIC, true, "The minimum amount of time between random events."));
+			this.AddConfig(new ConfigSetting("575_random_max", 300, SettingType.NUMERIC, true, "The maximum wait time between random events."));
+			this.AddConfig(new ConfigSetting("575_random_dur_min", 11, SettingType.NUMERIC, true, "The minimum amount of time a random event should last."));
+			this.AddConfig(new ConfigSetting("575_random_dur_max", 90, SettingType.NUMERIC, true, "The maximum amount of time a random event should last."));
 			this.AddEventHandlers(new EventsHandler(this), Priority.Normal);
-			//Timing.Init(this);
 			new Functions(this);
 			this.AddCommands(new string[] { "SCP575", "575" }, new SCP575Command());
 		}
